feat: add iwd support and improve installer and init
* tiny installer improvement

* wip iwd, installer improvements, init improvements

* improved strap

* removed duplicate from .gitignore

* docs: update dev branch cos yes

* removed duplicate y/n

* chore: remove duplicate newt

* fix(dbus-service): use /dev/console instead of /dev/tty1

* chore(.gitignore): wildcard kernel image gitignore patterns

* fix(make:vm): do not rebuild on make target vm

* fix: hostname fixing

* fix: drop spurious ISO prereq from installed-vm

+++ strap.py
@@ -1,22 +1,53 @@
 import os
+import subprocess
+
+"""
+Bootstrap packages into rootfs/filesystem. Basically a very simple version of Car.
+Requires Car to be installed and initialized.
+Usage:
+    echo "package" | python3 strap.py
+Prints an error message and exits 1 if package is not found.
+On success, prints the version of the package and adds it to rootfs/filesystem/etc/repro.car, then exits 0.
+"""
 
 packagelist = open("/etc/car/packagelist", "r").read()
+version = ""
 package = input()
-with open("rootfs/filesystem/etc/repro.car", "a") as f:
-    f.write(package + "\n")
-os.system("mkdir -p strap_packages")
+currently_at_package = False
+
+os.makedirs("strap_packages", exist_ok=True)
+os.makedirs("rootfs/filesystem/etc/car/saves", exist_ok=True)
+
 for i in packagelist.splitlines():
     if i.startswith(f"{package} - "):
+        print("installing " + package)
+        currently_at_package = True
         url = i.split(" - ")[1]
         print("downloading " + url)
-        if not os.path.exists("strap_packages/" + package + ".tar.zst"):
-            os.system("curl -# -L -o strap_packages/" + package + ".tar.zst " + url)
-        print("installing " + package)
-        os.system(
-            "tar -I zstd -xf strap_packages/"
-            + package
-            + ".tar.zst -C rootfs/filesystem --strip-components=1"
+        tarball = "strap_packages/" + package + ".tar.zst"
+        if not os.path.exists(tarball):
+            os.system("curl -# -L -o " + tarball + " " + url)
+
+        save_path = "rootfs/filesystem/etc/car/saves/" + package
+        result = subprocess.run(
+            f"tar -I 'zstd -T0' -xvf {tarball} -C rootfs/filesystem --strip-components=1 "
+            f"| sed 's|^[^/]*/||' | grep -v '/$' > {save_path}",
+            shell=True,
         )
+        if result.returncode != 0:
+            print(f"error: failed to unpack {tarball}")
+            exit(1)
+
+    elif currently_at_package:
+        if i.startswith("version "):
+            version = i.split(" ")[1]
+            print("version " + version)
+            with open("rootfs/filesystem/etc/repro.car", "a") as f:
+                f.write(package + "=" + version + "\n")
         break
+
+if not currently_at_package:
+    print("package not found")
+    exit(1)
 
 os.system("rm -f rootfs/filesystem/car")
